+++ choice_api/__init__.py
@@ -72,7 +72,7 @@ from .indicators import (
     cpr,
 )
 
-__version__ = "1.5.0"
+__version__ = "1.5.1"
 __all__ = [
     "ChoiceClient",
     "BASE_URL_OMNE",